[iOS] split viewmodel logic among main home and pets
