+++ iosApp/iosApp/Home/ViewModel/HomeViewModel.swift
@@ -1,0 +1,24 @@
+import Combine
+import shared
+
+class HomeViewModel: ObservableObject {
+    @Published private(set) var selectedOption = MenuOption.Pets
+    
+    @Published var genders = [
+        Gender_.female: true,
+        Gender_.male: true,
+    ]
+    
+    @Published var showMenu = false
+//    {
+//        didSet {
+//            if (!showMenu) {
+//                viewCreated()
+//            }
+//        }
+//    }
+    
+    func selectedOption(newOption: MenuOption) {
+        selectedOption = newOption
+    }
+}
